// pokemonApp/src/app/interfaces/pokemon.interface.ts
// Generated by https://quicktype.io

export interface PokemonInterface {
    id: number;
    location_area_encounters: string;
    name: string;
    order: number;
    sprites: Sprites;
    types: Type[];
}

export interface Species {
    name: string;
    url: string;
}

export interface Type {
    slot: number;
    type: Species;
}

export interface Sprites {
    back_default:       string;
    back_female:        string;
    back_shiny:         string;
    back_shiny_female:  string;
    front_default:      string;
    front_female:       string;
    front_shiny:        string;
    front_shiny_female: string;
}